+++ CHANGELOG.md
@@ -11,7 +11,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ### Changed
 - Scene Loading Setting asset uses Scriptable Settings Provider
-- Replace ScreenFaderCanvas -> ScreenFaderCanvas
+- Replace ScreenFaderCanvas -> CanvasScreenFader
 
 ### Removed
 - Animation Parameters package