+++ CHANGELOG.md
@@ -17,6 +17,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Animation Parameters package
 - ScreenFaderAnimator component
 - Static SceneManager class
+- SceneLoadingSettingsEditor
 
 ## [1.0.1] - 2021-01-14
 ### Changed