+++ CHANGELOG.md
@@ -6,6 +6,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 ### Added
+- Awaitable Coroutines package 1.0.0
 - Scriptable Settings Provider package 1.0.0
 - IScreenFader interface
 - AnimationScreenFader component and prefab